+++ src/pages/portfolio/portfolio.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PortfolioCard from "../components/portfolioCard";
+import PortfolioCard from "../../components/parts/portfolioCard/portfolioCard";
 import cards from "../../cards.json";
 import "./portfolio.css";
